+++ src/_pyorc/Reader.cpp
@@ -4,7 +4,7 @@
 #include "Reader.h"
 
 py::object
-ORCIterator::next()
+ORCFileLikeObject::next()
 {
     while (true) {
         if (batchItem == 0) {
@@ -24,8 +24,8 @@ ORCIterator::next()
     }
 }
 
-py::object
-ORCIterator::read(int64_t num)
+py::list
+ORCFileLikeObject::read(int64_t num)
 {
     int64_t i = 0;
     py::list res;
@@ -46,7 +46,7 @@ ORCIterator::read(int64_t num)
 }
 
 uint64_t
-ORCIterator::seek(int64_t row, uint16_t whence)
+ORCFileLikeObject::seek(int64_t row, uint16_t whence)
 {
     uint64_t start = 0;
     switch (whence) {
@@ -70,6 +70,158 @@ ORCIterator::seek(int64_t row, uint16_t whence)
     batchItem = 0;
     currentRow = rowReader->getRowNumber() - firstRowOfStripe;
     return currentRow;
+}
+
+const orc::Type*
+ORCFileLikeObject::findColumnType(const orc::Type* type, uint64_t columnIndex) const
+{
+    if (type->getColumnId() == columnIndex) {
+        return type;
+    } else {
+        for (size_t i = 0; i < type->getSubtypeCount(); ++i) {
+            auto* subtype = type->getSubtype(i);
+            if (subtype->getColumnId() <= columnIndex &&
+                subtype->getMaximumColumnId() >= columnIndex) {
+                return ORCFileLikeObject::findColumnType(subtype, columnIndex);
+            }
+        }
+        throw py::index_error("column not found");
+    }
+}
+
+py::object
+ORCFileLikeObject::convertTimestampMillis(int64_t millisec) const
+{
+    py::object idx(py::int_(static_cast<int>(orc::TIMESTAMP)));
+    py::object from_orc = convDict[idx].attr("from_orc");
+    int64_t seconds = millisec / 1000;
+    int64_t nanosecs = std::abs(millisec % 1000) * 1000 * 1000;
+    return from_orc(seconds, nanosecs);
+}
+
+py::dict
+ORCFileLikeObject::buildStatistics(const orc::Type* type,
+                                   const orc::ColumnStatistics* stats) const
+{
+    py::dict result;
+    int64_t typeKind = static_cast<int64_t>(type->getKind());
+    result["kind"] = typeKind;
+    result["has_null"] = py::cast(stats->hasNull());
+    result["number_of_values"] = py::cast(stats->getNumberOfValues());
+    switch (typeKind) {
+        case orc::BOOLEAN: {
+            auto* boolStat = dynamic_cast<const orc::BooleanColumnStatistics*>(stats);
+            if (boolStat->hasCount()) {
+                result["false_count"] = py::cast(boolStat->getFalseCount());
+                result["true_count"] = py::cast(boolStat->getTrueCount());
+            }
+            return result;
+        }
+        case orc::BYTE:
+        case orc::INT:
+        case orc::LONG:
+        case orc::SHORT: {
+            auto* intStat = dynamic_cast<const orc::IntegerColumnStatistics*>(stats);
+            if (intStat->hasMinimum()) {
+                result["minimum"] = py::cast(intStat->getMinimum());
+            }
+            if (intStat->hasMaximum()) {
+                result["maximum"] = py::cast(intStat->getMaximum());
+            }
+            if (intStat->hasSum()) {
+                result["sum"] = py::cast(intStat->getSum());
+            }
+            return result;
+        }
+        case orc::STRUCT:
+        case orc::MAP:
+        case orc::LIST:
+        case orc::UNION:
+            return result;
+        case orc::FLOAT:
+        case orc::DOUBLE: {
+            auto* doubleStat = dynamic_cast<const orc::DoubleColumnStatistics*>(stats);
+            if (doubleStat->hasMinimum()) {
+                result["minimum"] = py::cast(doubleStat->getMinimum());
+            }
+            if (doubleStat->hasMaximum()) {
+                result["maximum"] = py::cast(doubleStat->getMaximum());
+            }
+            if (doubleStat->hasSum()) {
+                result["sum"] = py::cast(doubleStat->getSum());
+            }
+            return result;
+        }
+        case orc::BINARY: {
+            auto* binaryStat = dynamic_cast<const orc::BinaryColumnStatistics*>(stats);
+            if (binaryStat->hasTotalLength()) {
+                result["total_length"] = py::cast(binaryStat->getTotalLength());
+            }
+            return result;
+        }
+        case orc::STRING:
+        case orc::CHAR:
+        case orc::VARCHAR: {
+            auto* strStat = dynamic_cast<const orc::StringColumnStatistics*>(stats);
+            if (strStat->hasMinimum()) {
+                result["minimum"] = py::cast(strStat->getMinimum());
+            }
+            if (strStat->hasMaximum()) {
+                result["maximum"] = py::cast(strStat->getMaximum());
+            }
+            if (strStat->hasTotalLength()) {
+                result["total_length"] = py::cast(strStat->getTotalLength());
+            }
+            return result;
+        }
+        case orc::DATE: {
+            auto* dateStat = dynamic_cast<const orc::DateColumnStatistics*>(stats);
+            py::object idx(py::int_(static_cast<int>(orc::DATE)));
+            py::object from_orc = convDict[idx].attr("from_orc");
+            if (dateStat->hasMinimum()) {
+                result["minimum"] = from_orc(dateStat->getMinimum());
+            }
+            if (dateStat->hasMaximum()) {
+                result["maximum"] = from_orc(dateStat->getMaximum());
+            }
+            return result;
+        }
+        case orc::TIMESTAMP: {
+            auto* timeStat = dynamic_cast<const orc::TimestampColumnStatistics*>(stats);
+            if (timeStat->hasMinimum()) {
+                result["minimum"] = convertTimestampMillis(timeStat->getMinimum());
+            }
+            if (timeStat->hasMaximum()) {
+                result["maximum"] = convertTimestampMillis(timeStat->getMaximum());
+            }
+            if (timeStat->hasLowerBound()) {
+                result["lower_bound"] =
+                  convertTimestampMillis(timeStat->getLowerBound());
+            }
+            if (timeStat->hasUpperBound()) {
+                result["upper_bound"] =
+                  convertTimestampMillis(timeStat->getUpperBound());
+            }
+            return result;
+        }
+        case orc::DECIMAL: {
+            auto* decStat = dynamic_cast<const orc::DecimalColumnStatistics*>(stats);
+            py::object idx(py::int_(static_cast<int>(orc::DECIMAL)));
+            py::object from_orc = convDict[idx].attr("from_orc");
+            if (decStat->hasMinimum()) {
+                result["minimum"] = from_orc(decStat->getMinimum().toString());
+            }
+            if (decStat->hasMaximum()) {
+                result["maximum"] = from_orc(decStat->getMaximum().toString());
+            }
+            if (decStat->hasSum()) {
+                result["sum"] = from_orc(decStat->getSum().toString());
+            }
+            return result;
+        }
+        default:
+            return result;
+    }
 }
 
 Reader::Reader(py::object fileo,
@@ -97,19 +249,18 @@ Reader::Reader(py::object fileo,
     if (conv.is(py::none())) {
         py::dict defaultConv =
           py::module::import("pyorc.converters").attr("DEFAULT_CONVERTERS");
-        converters = py::dict(defaultConv);
+        convDict = py::dict(defaultConv);
     } else {
-        converters = conv;
+        convDict = conv;
     }
     reader = createReader(
       std::unique_ptr<orc::InputStream>(new PyORCInputStream(fileo)), readerOpts);
-    typeDesc = std::make_unique<TypeDescription>(reader->getType());
     try {
         batchSize = batch_size;
         rowReader = reader->createRowReader(rowReaderOpts);
         batch = rowReader->createRowBatch(batchSize);
         converter =
-          createConverter(&rowReader->getSelectedType(), structKind, converters);
+          createConverter(&rowReader->getSelectedType(), structKind, convDict);
     } catch (orc::ParseError& err) {
         throw py::value_error(err.what());
     }
@@ -136,10 +287,32 @@ Reader::readStripe(uint64_t idx)
     return Stripe(*this, idx, reader->getStripe(idx));
 }
 
-TypeDescription&
+TypeDescription
 Reader::schema()
 {
-    return *typeDesc;
+    return TypeDescription(reader->getType());
+}
+
+TypeDescription
+Reader::selectedSchema()
+{
+    return TypeDescription(rowReader->getSelectedType());
+}
+
+py::tuple
+Reader::statistics(uint64_t columnIndex)
+{
+    try {
+        py::tuple result = py::tuple(1);
+        std::unique_ptr<orc::ColumnStatistics> stats =
+          reader->getColumnStatistics(columnIndex);
+        result[0] = this->buildStatistics(
+          this->findColumnType(&rowReader->getSelectedType(), columnIndex),
+          stats.get());
+        return result;
+    } catch (std::logic_error& err) {
+        throw py::index_error(err.what());
+    }
 }
 
 Stripe::Stripe(const Reader& reader_,
@@ -151,17 +324,18 @@ Stripe::Stripe(const Reader& reader_,
     currentRow = 0;
     stripeIndex = idx;
     stripeInfo = std::move(stripe);
+    convDict = reader.getConverterDict();
     rowReaderOpts = reader.getRowReaderOptions();
     rowReaderOpts =
       rowReaderOpts.range(stripeInfo->getOffset(), stripeInfo->getLength());
     rowReader = reader.getORCReader().createRowReader(rowReaderOpts);
     batch = rowReader->createRowBatch(reader.getBatchSize());
-    converter = createConverter(
-      &rowReader->getSelectedType(), reader.getStructKind(), reader.getConverters());
+    converter =
+      createConverter(&rowReader->getSelectedType(), reader.getStructKind(), convDict);
     firstRowOfStripe = rowReader->getRowNumber() + 1;
 }
 
-py::object
+py::tuple
 Stripe::bloomFilterColumns()
 {
     int64_t idx = 0;
@@ -192,6 +366,26 @@ uint64_t
 Stripe::offset() const
 {
     return stripeInfo->getOffset();
+}
+
+py::tuple
+Stripe::statistics(uint64_t columnIndex)
+{
+    if (columnIndex < 0 ||
+        columnIndex > rowReader->getSelectedType().getMaximumColumnId()) {
+        throw py::index_error("column index out of range");
+    }
+    std::unique_ptr<orc::StripeStatistics> stripeStats =
+      reader.getORCReader().getStripeStatistics(stripeIndex);
+    uint32_t num = stripeStats->getNumberOfRowIndexStats(columnIndex);
+    py::tuple result = py::tuple(num);
+    for (uint32_t i = 0; i < num; ++i) {
+        const orc::ColumnStatistics* stats =
+          stripeStats->getRowIndexStatistics(columnIndex, i);
+        result[i] = this->buildStatistics(
+          this->findColumnType(&rowReader->getSelectedType(), columnIndex), stats);
+    }
+    return result;
 }
 
 std::string